+++ packages/bluesky_cli/lib/src/command/notification/notification_count.dart
@@ -1,0 +1,32 @@
+// Copyright 2023 Kato Shinya. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+import 'package:xrpc/xrpc.dart' as xrpc;
+
+import '../query_command.dart';
+
+/// The command for `app.bsky.notification.getCount`.
+class NotificationCountCommand extends QueryCommand {
+  /// Returns the new instance of [NotificationCountCommand].
+  NotificationCountCommand();
+
+  @override
+  final String name = 'notification-count';
+
+  @override
+  final String description =
+      'Show the count of notification of authenticated user.';
+
+  @override
+  final String invocation = 'bsky notification-count';
+
+  @override
+  xrpc.NSID get methodId => xrpc.NSID.create(
+        'notification.bsky.app',
+        'getCount',
+      );
+
+  @override
+  Map<String, dynamic>? get parameters => null;
+}